requirements.txt
selenium==4.9.1
beautifulsoup4==4.12.2
webdriver_manager==4.0.0
lxml==4.9.3
aiohttp==3.8.1
async-timeout==4.0.1
prefect-dask==0.2.6
prefect-aws==0.4.9